add notification component to app.tsx
